PERF: improve iloc list indexing
Author: Joris Van den Bossche <jorisvandenbossche@gmail.com>

Closes #15504 from jorisvandenbossche/perf-iloc-list and squashes the following commits:

bf54a0b [Joris Van den Bossche] TST: edit test_take to preserve original dtype
74d45ae [Joris Van den Bossche] add whatsnew
3e537b6 [Joris Van den Bossche] small clean-up
6d2705c [Joris Van den Bossche] take method: only validate kwargs if there are kwargs
aacbaa8 [Joris Van den Bossche] PERF: improve iloc list indexing

+++ pandas/tests/test_generic.py
@@ -1870,7 +1870,7 @@ class TestNDFrame(tm.TestCase):
                   tm.makeObjectSeries()]:
             out = s.take(indices)
             expected = Series(data=s.values.take(indices),
-                              index=s.index.take(indices))
+                              index=s.index.take(indices), dtype=s.dtype)
             tm.assert_series_equal(out, expected)
         for df in [tm.makeTimeDataFrame()]:
             out = df.take(indices)
